+++ internal/pkg/validate/doc.go
@@ -1,2 +1,0 @@
-// Package validate contains validation functions.
-package validate