+++ pkg/tinypng/errors.go
@@ -1,12 +1,25 @@
 package tinypng
 
-import "strings"
+import (
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"strings"
+)
+
+// Error is a special type for package-specific errors.
+type Error uint8
+
+// Package-specific error constants.
+const (
+	ErrTooManyRequests Error = iota + 1 // too many requests (limit has been exceeded)
+	ErrUnauthorized                     // unauthorized (invalid credentials)
+	ErrBadRequest                       // bad request (empty file or wrong format)
+)
 
 // Package-specific errors prefix.
 const errorsPrefix = "tinypng.com:"
-
-// Special type for package-specific errors.
-type Error uint8
 
 // Error returns error in a string representation.
 func (err Error) Error() string {
@@ -32,9 +45,26 @@ func (err Error) Error() string {
 	return buf.String()
 }
 
-// Package-specific error constants.
-const (
-	ErrTooManyRequests Error = iota + 1 // too many requests (limit has been exceeded)
-	ErrUnauthorized                     // unauthorized (invalid credentials)
-	ErrBadRequest                       // bad request (empty file or wrong format)
-)
+func newError(message string) error {
+	return fmt.Errorf(errorsPrefix + " " + message)
+}
+
+func newErrorf(format string, args ...any) error {
+	return errors.New(errorsPrefix + " " + fmt.Errorf(format, args...).Error())
+}
+
+// parseRemoteError reads HTTP response content as a JSON-string, parse them and converts into go-error.
+//
+// This function should never return nil!
+func parseRemoteError(content io.Reader) error {
+	var e struct {
+		Error   string `json:"error"`
+		Message string `json:"message"`
+	}
+
+	if err := json.NewDecoder(content).Decode(&e); err != nil {
+		return newErrorf("error decoding failed: %w", err)
+	}
+
+	return newErrorf("%s (%s)", e.Error, strings.Trim(e.Message, ". "))
+}